backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==11.2.1
coreapi==2.3.3
python-dotenv==1.0.0
django-filter==23.5
pyyaml==6.0.1
drf-yasg==1.21.7
psycopg2-binary==2.9.10
dj-database-url==2.1.0
gunicorn==21.2.0
